Bug 824247 - Part 1: Fix common source of warning spam (dom). r=smaug
UltraBlame original commit: 0e626106842fae726f4c2408657c0da299538794

+++ content/base/src/nsDOMBlobBuilder.h
@@ -125,7 +125,8 @@ protected:
     }
 
     
-    CheckedUint32 bufferLen = NS_MAX<uint32_t>(mDataBufferLen, 1);
+    CheckedUint32 bufferLen =
+      NS_MAX<uint32_t>(static_cast<uint32_t>(mDataBufferLen), 1);
     while (bufferLen.isValid() && bufferLen.value() < mDataLen + aSize)
       bufferLen *= 2;
